% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{get_deps}
\alias{get_deps}
\title{Get dependencies}
\usage{
get_deps(file)
}
\arguments{
\item{file}{a string character with the path for a R script file}
}
\value{
a character vector with dependencies
}
\description{
Get dependencies
}
